added the reconstructor class.
It ain't much, but it's honest work.

still would fail miserably, if single events start missing due to detector noise. Also we still need to increase the time dt between the creation of particles, to make it one particle between z = 30 and z = 50 at all time, to make the Code run. So still pretty shit overall and still in work